requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
Django==5.2.3
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pygame==2.6.1
Pygments==2.19.2
PyJWT==1.7.1
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
